Move username generator under the user module
